// frontend/src/plugins/metal/preparations/meropenem/models/SamplePreparationMeropenem.ts
export interface SamplePreparationMeropenemStep {
    name: string;
    value1?: string;
    unit1?: string;
    value2?: string;
    unit2?: string;
    value3?: string;
    unit3?: string;
    logBookID?: string;
}

export interface SamplePreparationMeropenem {
    id: number;
    label: string;
    steps: SamplePreparationMeropenemStep[];
}